requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
Pillow==8.2.0
protobuf==3.17.1
six==1.16.0
Werkzeug==2.0.1
